multimodal-analysis/requirements.txt
# Gemini SDK
google-generativeai>=0.8.0

# LangChain
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-chroma>=0.1.0

# 向量存储
chromadb>=0.5.0

# 图像处理
Pillow>=10.0.0

# 视频处理
moviepy>=1.0.3

# 音频处理
pydub>=0.25.0

# Web 框架
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12

# 数据模型
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 终端 UI
rich>=13.0.0

# 环境变量
python-dotenv>=1.0.0

# 类型支持
typing-extensions>=4.0.0
